+++ src/lib/utils.ts
@@ -1,10 +1,10 @@
-export function byteArrayToHex(bytes: Uint8Array): string {
-  const buf = new Buffer(bytes);
-  return buf.toString('hex');
+export function hexToBuf(str: string): Buffer {
+  return Buffer.from(str.replace(/^0x/i, ''), 'hex');
 }
 
-export function hexToBuf(str: string): Buffer {
-  return Buffer.from(str.replace(/0x/, ''), 'hex');
+export function bufToHex(buf: Uint8Array | Buffer | ArrayBuffer): string {
+  const _buf = Buffer.from(buf);
+  return _buf.toString('hex');
 }
 
 export function equalByteArrays(